# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Food diary API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /foods/parse:
    post:
      operationId: parseFood
      tags: [foods]
      summary: Use AI to parse a natural-language food description into structured nutrition data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParseFoodInput"
      responses:
        "200":
          description: Parsed nutrition data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParsedFood"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: AI service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /foods:
    post:
      operationId: createFood
      tags: [foods]
      summary: Save a custom food to the local database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFoodInput"
      responses:
        "201":
          description: Food created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Food"
  /foods/search:
    get:
      operationId: searchFoods
      tags: [foods]
      summary: Search locally cached foods
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of matching foods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Food"
  /foods/lookup:
    post:
      operationId: lookupFood
      tags: [foods]
      summary: Look up food by barcode from Open Food Facts and save to local cache
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodLookupInput"
      responses:
        "200":
          description: Found and returned food
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Food"
        "404":
          description: Product not found in Open Food Facts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /foods/history:
    get:
      operationId: getFoodHistory
      tags: [foods]
      summary: Get the authenticated user's recent food scan history
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Paginated scan history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodHistoryResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Food:
      type: object
      properties:
        id:
          type: number
        barcode:
          type: ["string", "null"]
        name:
          type: string
        brand:
          type: ["string", "null"]
        servingSize:
          type: ["string", "null"]
        kcalPerServing:
          type: ["number", "null"]
        fatPerServing:
          type: ["number", "null"]
        saturatedFatPerServing:
          type: ["number", "null"]
        carbsPerServing:
          type: ["number", "null"]
        sugarsPerServing:
          type: ["number", "null"]
        fiberPerServing:
          type: ["number", "null"]
        proteinPerServing:
          type: ["number", "null"]
        saltPerServing:
          type: ["number", "null"]
        sodiumPerServing:
          type: ["number", "null"]
        potassiumPerServing:
          type: ["number", "null"]
        cholesterolPerServing:
          type: ["number", "null"]
        nutriscoreGrade:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
      required:
        - id
        - name
    ParseFoodInput:
      type: object
      properties:
        query:
          type: string
          minLength: 1
      required:
        - query
    ParsedFood:
      type: object
      properties:
        foodName:
          type: string
        quantity:
          type: number
        unit:
          type: string
        calories:
          type: number
        proteinG:
          type: number
        carbsG:
          type: number
        fatG:
          type: number
      required:
        - foodName
        - quantity
        - unit
        - calories
        - proteinG
        - carbsG
        - fatG
    CreateFoodInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        servingSize:
          type: ["string", "null"]
        kcalPerServing:
          type: ["number", "null"]
        fatPerServing:
          type: ["number", "null"]
        saturatedFatPerServing:
          type: ["number", "null"]
        carbsPerServing:
          type: ["number", "null"]
        sugarsPerServing:
          type: ["number", "null"]
        fiberPerServing:
          type: ["number", "null"]
        proteinPerServing:
          type: ["number", "null"]
        saltPerServing:
          type: ["number", "null"]
        sodiumPerServing:
          type: ["number", "null"]
        potassiumPerServing:
          type: ["number", "null"]
        cholesterolPerServing:
          type: ["number", "null"]
      required:
        - name
    FoodLookupInput:
      type: object
      properties:
        barcode:
          type: string
      required:
        - barcode
    FoodHistoryEntry:
      type: object
      properties:
        id:
          type: number
        foodId:
          type: number
        scannedAt:
          type: string
        food:
          $ref: "#/components/schemas/Food"
      required:
        - id
        - foodId
        - scannedAt
        - food
    FoodHistoryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FoodHistoryEntry"
        pagination:
          type: object
          properties:
            limit:
              type: number
            offset:
              type: number
            hasMore:
              type: boolean
          required:
            - limit
            - offset
            - hasMore
      required:
        - items
        - pagination
...