# lib/api-zod/src/generated/types/foodHistoryEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Food diary API
 * OpenAPI spec version: 0.1.0
 */
import type { Food } from './food';

export interface FoodHistoryEntry {
  id: number;
  foodId: number;
  scannedAt: string;
  food: Food;
}