# lib/api-zod/src/generated/types/foodHistoryResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Food diary API
 * OpenAPI spec version: 0.1.0
 */
import type { FoodHistoryEntry } from './foodHistoryEntry';
import type { FoodHistoryResponsePagination } from './foodHistoryResponsePagination';

export interface FoodHistoryResponse {
  items: FoodHistoryEntry[];
  pagination: FoodHistoryResponsePagination;
}